% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conn-accessinfo.R
\name{AccessInfo}
\alias{AccessInfo}
\alias{CheckDB}
\alias{ConnString}
\alias{Databases}
\alias{ValidDB}
\title{Access Information}
\usage{
ConnString(db = NULL)

AccessInfo(db = NULL)

ValidDB()

CheckDB(db = NULL)

Databases()
}
\arguments{
\item{db}{A character value representing the name of a database}
}
\description{
These functions are intended for developers that need access to the connection
environment as they develop data products.
}
\section{Functions}{
\itemize{
\item \code{ConnString}: Returns the connection string associated with
a configured database

\item \code{AccessInfo}: See the access params for configured databases

\item \code{ValidDB}: See the current set of configured databases

\item \code{CheckDB}: A function to check the argument against a list of
valid DBs. Returns error on a none match. Performs approximate match.

\item \code{Databases}: 
}}
\section{Note}{

Any database where access to managing connections is required must be configured prior.
Future functionality will include the ability to auto-configurate a database.
}

