% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitAuth.R
\name{BackendAuth}
\alias{BackendAuth}
\alias{CheckOrg}
\alias{Databases}
\alias{FileMove}
\alias{GitApp}
\alias{github_api}
\alias{github_pat}
\title{Github Authentication}
\usage{
GitApp()

github_api(path, gtoken = NULL, json = TRUE)

CheckOrg(orgname = "intusurg")

Databases(bUpdate = FALSE)

FileMove(fpath = NULL, destn = NULL, overwrite = FALSE, makedir = TRUE)

github_pat()
}
\arguments{
\item{path}{The API path}

\item{gtoken}{Additional headers for API call}

\item{json}{A boolean to indicate whether to check response for json format. Default = TRUE}

\item{orgname}{Character vector of length 1 that represents the name of the organization}

\item{bUpdate}{A boolean representing whether to update the local access file}

\item{fpath}{File to move}

\item{destn}{new location to move to}

\item{overwrite}{A boolean indicating whether to overwrite existing file at destination}

\item{makedir}{A boolean indicating whether to create a new directory for destination file
if one did not previously exist (default = TRUE)}
}
\description{
Functions to enable backend github authentication
}
\section{Functions}{
\itemize{
\item \code{GitApp}: returns github app key

\item \code{github_api}: A basic function to interact with Github API.
Proper credits to Hadley Wickham coming.

\item \code{CheckOrg}: A function that checks whether the authenticated github user
is a member of the given organization

\item \code{Databases}: An updated version of the prior function to pull access info.
This function pulls access info from remote location and stores it locally.
It can also update local file

\item \code{FileMove}: A better function to move files. Move later

\item \code{github_pat}: A function to retrieve the locally stored access token
}}

