% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conn-attributes.R
\name{ConnectionAttributes}
\alias{ConnStatus}
\alias{ConnectionAttributes}
\alias{ListConnAttr}
\alias{SeeClosedConns}
\alias{SeeConn}
\alias{SeeOpenConns}
\title{Connection Attribute Information}
\usage{
SeeConn(db = NULL, what = NULL)

SeeOpenConns()

SeeClosedConns()

ConnStatus()

ListConnAttr()
}
\arguments{
\item{db}{A character value representing the name of a database whose connection's attributes need to be accessed}

\item{what}{A character value representing the attribute name to access}
}
\description{
Resources to manage connections, to test the status of connections, as well as functions to get details (attributes)
about a connection object that exists within the connection pool.
}
\details{
As a developer creating a custom query function, you may need the ability to
know, for example, if a connection is open and how long it's been open. These are known
as attributes and all existing connection objects have the following set of attributes:

\itemize{
 \item {\code{Database}}{Name of database}
 \item {\code{Status}}{Open/Closed}
 \item {\code{Initiator}}{The first opener of the connection}
 \item {\code{Opener}}{Who opened a closed connection}
 \item {\code{Closer}}{Who closed an open connection}
 \item {\code{Requestor}}{Who requested a connection (e.g. tried to open an already
         open connection)}
 \item {\code{TimeInitiated}}{When was the connection first opened (timestamp resets
         when connection is closed)}
 \item {\code{TimeOpened}}{When did the connection's status change from Closed to Open}
 \item {\code{TimeClosed}}{When did the connection's status change from Open to Closed}
 \item {\code{TimeRequested}}{When was the connection was last requested (timestamp)}
 \item {\code{DurSinceInit}}{Duration since connection was initiated}
 \item {\code{DurSinceRequest}}{Duration since a connection was last requested}
 \item {\code{DurationOpen}}{Duration that connection has been open}
 \item {\code{DurationClosed}}{Duration that connection has been closed}
 \item {\code{AccessCount}}{Number of times the connection has been requested/opened since init}
}
}
\section{Functions}{
\itemize{
\item \code{SeeConn}: A function to "see" a particular connection objects attributes.

\item \code{SeeOpenConns}: See attributes for all open connections

\item \code{SeeClosedConns}: See attributes for all closed connections

\item \code{ConnStatus}: Prints status of all connection in the pool

\item \code{ListConnAttr}: A function that provides a quick lookup of
all attribute names that an object will have
}}
\examples{
\dontrun{
# Common developer helper functions #
db <- "a_database_name"
CheckDB(db)     # trys to find a matching database name, returns error if not
AccessInfo(db)  # access info for specified db
ValidDB()       # What DBs are configured in the package?
ConnString(db)  # Connection string for db?

# Opening/Closing Connections #
OpenDB(db)      # Returns 1 on success
CloseDB(db)     # Returns 1 on success
CloseDB(db)     # Warning if closing db that is already closed
CloseDB("blah") # Error if db doesn't exist

# Manage connections #
ConnExists(db)  # Returns TRUE for any open/closed connections
ConnPool()      # Returns all the connection (names) in the pool
ConnStatus()    # Returns a table showing status for all configured dbs

# Use should be rare to not at all #
GetConn(db)     # Retrieve connection object
Clean(db)       # Removes connection from pool (usually because of error)
Clean()         # Removes ALL connections from pool (usually because of error)
}
}

