% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-debughandling.R
\docType{data}
\name{DebugHelp}
\alias{.CallStack}
\alias{.Caller}
\alias{DebugHelp}
\alias{PrintSqlError}
\alias{Timer}
\alias{time.env}
\alias{xtimetaken}
\title{Internal Dev Testing and Debugging Functions}
\format{An object of class \code{environment} of length 0.}
\usage{
time.env

PrintSqlError(msg = NULL, friendly = NULL, nFrame = NULL)

xtimetaken(time)

Timer(START = TRUE, print = FALSE)

.CallStack(nFrame)

.Caller(nFrame = NULL)
}
\arguments{
\item{msg}{A character string representing the message to display within the banner}

\item{friendly}{A friendly description of the error message to be displayed}

\item{nFrame}{A numeric value representing an index for a frame in the function callstack}

\item{time}{A date/time object from which point to calculate duration since.}

\item{START}{A boolean value representing whether to start the timer or end the timer. START = FALSE
will display the timer information}

\item{print}{A boolean indicating whether to print output on the console in addition to returning a numeric
value representing the duration in seconds}
}
\description{
These are functions that are not exposed to users. They are internal functions designed to assist
with debugging errors related to tracing and identifying connections.
}
\section{Functions}{
\itemize{
\item \code{PrintSqlError}: A helper function to display sql errors that occur during a query

\item \code{xtimetaken}: A helper function to assist timing operations

\item \code{Timer}: A helper function to assist timing operations

\item \code{.CallStack}: A helper function to extract the relevant and helpful portion of the function callstack

\item \code{.Caller}: A function to parse the function call stack and identify a caller at the current postion
or at the position identified by "nFrame"
}}
\section{Internal Use Only}{

Internal functions not intended to be used by developers.
}
\keyword{datasets}

