% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conn-openclose.R
\name{OpenDB}
\alias{.odbcClose}
\alias{.odbcOpen}
\alias{.odbcReOpen}
\alias{CloseDB}
\alias{OpenDB}
\title{Open and Close Connections}
\usage{
OpenDB(db = NULL, verbose = F)

CloseDB(db = NULL, verbose = F)

.odbcOpen(db)

.odbcClose(db)

.odbcReOpen(db)
}
\arguments{
\item{db}{A character value naming the database to use}

\item{verbose}{A boolean indicating whether to print a status report on the console}
}
\description{
A number of functions to handle connections. These functions are intended for developers that need
access to the connection environment as they develop data products. All functions within the morphr package
utilize these functions. If you are a developer would like more details, see the examples below or visit
the following tutorial pages:
}
\details{
Any database where access to managing connections is required must be configured prior.

A connection object can have multiple states:
\itemize{
 \item Open/closed
 \item Not Open/Not Closed
}

The latter two ("Not Open" and "Not Closed") refer to states where a connection object may exist but has timeout,
or when the connection object does not exist at all. For example, the first time a connection is initialized, it
becomes "Open". Prior to that, it is not open or closed. After a request is received to close the connection, the
status is changed to "closed". From there on, the connection object exists in the pool until it is garbage collected
or explicitly removed.
}
\section{Functions}{
\itemize{
\item \code{OpenDB}: Open connection to a given database. Returns Null

\item \code{CloseDB}: Close a connection to given db. Returns Null

\item \code{.odbcOpen}: Internal RODBC wrapper

\item \code{.odbcClose}: Internal RODBC wrapper

\item \code{.odbcReOpen}: Internal RODBC wrapper
}}
\section{Internal Functions}{

These internal functions are wrappers for package:RODBC functions. The wrappers are there primarily for custom
error handling and are not intended to be exposed to developers or users.
}

