% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QuerySauce.R
\name{QuerySauce}
\alias{QuerySauce}
\alias{WhereSauce}
\title{The Sauce for Dishing Data}
\usage{
QuerySauce(tbl = NULL, top = NULL, cols = NULL, where = NULL)

WhereSauce(...)
}
\arguments{
\item{tbl}{A character string that represents the sql tablename to query}

\item{top}{A numeric value indicating whether to return the top N rows from the
query execution}

\item{cols}{An optional character vector specifying columns to return. If not
provided, all columns will be returned}

\item{where}{The output from a call to WhereSauce. A character string representing
     a valid where clause}
}
\description{
The function \code{QuerySauce} and it's helper \code{WhereSauce} provide the
the ability to effortlessly build a valid sql query string
}
\details{
FILL IN DETAIL HERE. See example workflows
}
\section{Functions}{
\itemize{
\item \code{QuerySauce}: A function to build the query

\item \code{WhereSauce}: A function to the WHERE clause that becomes an input into
\code{QuerySauce}
}}
\examples{

### Terminology

### Example workflow

### Operators for the where sauce


#----------------------------------------------------------------------------
# Work-flow to serve the data dish:
#
#               Where sauce --> Query sauce --> Data dish!
#
# STEP 1: Name the configured database
# STEP 2: Name the table to query
# STEP 3: Get the WHERE sauce (optional ingredient for the query sauce)
# STEP 4: Get the Query sauce (main ingredient for the data dish)
#----------------------------------------------------------------------------

db <- "RemoteFE"
tbl <- "SFDC.dbo.PROCEDURES"

#----------------------------------------------------------------------------
# Get procedures of specified subjects that occurred since 2015
#----------------------------------------------------------------------------

# Ingredients for the WHERE sauce
psub <- c("HPB", "dVP", "dVC", "dVL")       # procedure subjects
yr <- 2015                                  # year condition


wh    <- WhereSauce(ProcedureSubject == psub & Year(CreateDate) > yr)
query <- QuerySauce(tbl, top = 1000, wh)


OpenDB(db)                  # Open connection
DT <- xQuery(db, query)     # Run query
CloseDB(db)                 # Close connection

}

