% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conn-access_attributes.R
\name{AccessInfo}
\alias{AccessCount}
\alias{AccessInfo}
\alias{CloseConnStatus}
\alias{CloseDuration}
\alias{ClosedBy}
\alias{ConnAttr}
\alias{ConnStatus}
\alias{DurSinceLastRequest}
\alias{IsActive}
\alias{IsClosed}
\alias{IsOpen}
\alias{OpenConnStatus}
\alias{OpenDuration}
\alias{OpenedBy}
\alias{ReOpenedBy}
\alias{RequestedBy}
\alias{TimeClosed}
\alias{TimeLastRequest}
\alias{TimeOpened}
\title{Connection Object Information}
\usage{
AccessInfo(db)

ConnStatus(db = NULL, nFrame = NULL)

OpenConnStatus(db = NULL)

CloseConnStatus(db = NULL)

ConnAttr(db = NULL)

IsActive(db = NULL)

TimeClosed(db = NULL)

TimeOpened(db = NULL)

TimeLastRequest(db = NULL)

IsOpen(db = NULL)

IsClosed(db = NULL)

OpenedBy(db = NULL)

ReOpenedBy(db = NULL)

ClosedBy(db = NULL)

RequestedBy(db = NULL)

OpenDuration(db = NULL)

DurSinceLastRequest(db = NULL)

CloseDuration(db = NULL)

AccessCount(db = NULL)
}
\arguments{
\item{db}{A character value naming the database to use}

\item{nFrame}{An internally used parameter to provide the developer a traceback of where a connection error occured}
}
\description{
Functions to get more granular details about a connection object that exists within the connection pool.
For more general connection handling functions
}
\details{
These functions are convenience wrappers that retrieve attribute information from a connection object that exists.
The functions presented here are intended for developers and adminstrators and are likely less common requests. For additional
connection management functions that are perhaps more common, see
}
\section{Functions}{
\itemize{
\item \code{AccessInfo}: Prints information about access parameters to specified database. Returns NULL

\item \code{ConnStatus}: Prints verbose information on the current status of specified database. Returns Null

\item \code{OpenConnStatus}: Convenience function to provide quick status for an open db

\item \code{CloseConnStatus}: Convenience function to provide quick status for a closed db

\item \code{ConnAttr}: Convenience function to provide quick status for any db

\item \code{IsActive}: A function can be "inactive" which means it has never been initiated and neither open nor closed

\item \code{TimeClosed}: Returns timestamp of when connection was last closed

\item \code{TimeOpened}: Returns timestamp of when connection was first opened

\item \code{TimeLastRequest}: Returns timestamp of when connection was last requested (last opened)

\item \code{IsOpen}: Returns BOOL indicating whether status is "Open"

\item \code{IsClosed}: Returns BOOL indicating whether status is "Closed"

\item \code{OpenedBy}: Returns a string indicating the custom query function or user that make the FIRST opened connection to
specified database

\item \code{ReOpenedBy}: Returns a string indicating the custom query function or user that make the LAST opened connection to
specified database

\item \code{ClosedBy}: Returns a string indicating the custom query function or user most recently closed a connection
to the specified database

\item \code{RequestedBy}: Returns a string indicating the custom query function or user that sent a request to make a connection
to the specified database. A "request" simply means that an "open" request was made to a database that was already open.

\item \code{OpenDuration}: Returns the duration in seconds that an open connection has existed for a specified db

\item \code{DurSinceLastRequest}: Returns the duration in seconds since the last request was made to a specifed database

\item \code{CloseDuration}: Returns the duration in seconds since the status of the specifed database was changed to "close"

\item \code{AccessCount}: Returns a numeric that represents the count of requests that were made to a specified db
}}

