% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Internal_DebugHelpers.R
\name{DebugInfo}
\alias{.CallStack}
\alias{.Caller}
\alias{.PrintInfo}
\alias{DebugInfo}
\alias{PrintMessage}
\alias{PrintSqlError}
\alias{Timer}
\alias{timetaken}
\title{Internal Dev Testing and Debugging Functions}
\usage{
DebugInfo(msg = NULL, nFrame = NULL)

PrintSqlError(msg = NULL, friendly = NULL, nFrame = NULL)

timetaken(time)

Timer(START = TRUE, print = FALSE)

.CallStack(nFrame)

.Caller(nFrame = NULL)

.PrintInfo(ll, label, title)

PrintMessage(msg, sym = "#", content = NULL)
}
\arguments{
\item{msg}{A message to be displayed with an error}

\item{nFrame}{A numeric value representing an index for a frame in the function callstack}

\item{friendly}{A friendly description of the error message to be displayed}

\item{time}{A date/time object from which point to calculate duration since.}

\item{START}{A boolean value representing whether to start the timer or end the timer. START = FALSE
will display the timer information}

\item{ll}{A list with objects to print information about}

\item{label}{A character string used to label the information about the contents of "ll"}

\item{title}{A character string to place as a header title to the information printed on the console}

\item{sym}{A character symbol used to create the banner}

\item{content}{An optional object to print below the banner}

\item{msg}{A character string representing the message to display within the banner}
}
\description{
These are functions that are not exposed to users. They are internal functions designed to assist
with debugging errors related to tracing and identifying connections.
}
\section{Functions}{
\itemize{
\item \code{DebugInfo}: A function to customize errors related to connections

\item \code{PrintSqlError}: A helper function to display sql errors that occur during a query

\item \code{timetaken}: A helper function to assist timing operations

\item \code{Timer}: A helper function to assist timing operations

\item \code{.CallStack}: A helper function to extract the relevant and helpful portion of the function callstack

\item \code{.Caller}: A function to parse the function call stack and identify a caller at the current postion
or at the position identified by "nFrame"

\item \code{.PrintInfo}: A helper function and wrapper around "str" to print info to the console

\item \code{PrintMessage}: A function to elegantly print a notification banner on the console
}}
\section{Internal Use Only}{

 Internal functions not intended to be used by developers. To open and close connections to a preconfigured database,
 use: \code{\link{OpenDB}} or \code{\link{CloseDB}}.
}

