% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExecuteQuery.R
\name{xQuery}
\alias{xQuery}
\title{Execute a sql query on the database}
\usage{
xQuery(db, query, time = TRUE)
}
\arguments{
\item{db}{A database to send a query to and retrieve data from}

\item{query}{A valid SQL query string}

\item{time}{A boolean indicating wether to display the query's time to completion}
}
\value{
A data.table with query results
}
\description{
This function is a wrapper around \code{RODBC::sqlQuery} that provides some convenience as it interacts with the
connection handlers in this package. Additionally, it provides helpful errors, a tracing system, and various other checks
}

