% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn-manage_internal.R
\docType{data}
\name{InternalConnHandling}
\alias{InternalConnHandling}
\alias{closeConn}
\alias{cn.env}
\title{ConnectionHandlers}
\format{An object of class \code{environment} of length 0.}
\usage{
cn.env

closeConn(db, caller, ts)
}
\arguments{
\item{db}{internal database name}

\item{caller}{internal caller}

\item{ts}{internal timestamp}
}
\description{
Internal functions to get/set connection objects and attributes
}
\section{Functions}{
\itemize{
\item \code{closeConn}: internal function to close an open connection
}}
\keyword{datasets}

