% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conn-manage_internal.R
\docType{data}
\name{InternalConnHandling}
\alias{InternalConnHandling}
\alias{closeConn}
\alias{cn.env}
\title{ConnectionHandlers}
\format{\preformatted{<environment: 0x00000000071f5fc8> 
}}
\usage{
cn.env

closeConn(db, caller, ts)
}
\arguments{
\item{db}{}

\item{caller}{}

\item{ts}{}
}
\description{
Internal functions to get/set connection objects and attributes
}
\section{Functions}{
\itemize{
\item \code{closeConn}: 
}}
\keyword{datasets}

