% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conn-handling_dbconnect.R
\docType{data}
\name{cn.env}
\alias{CheckDB}
\alias{Clean}
\alias{CleanAll}
\alias{ConnString}
\alias{Databases}
\alias{GetConn}
\alias{SetConn}
\alias{cn.env}
\title{ConnectionHandlers}
\format{\preformatted{<environment: 0x0000000007e8dcc0> 
}}
\usage{
cn.env

GetConn(db)

SetConn(object = NULL, name = NULL, status = NULL, openedby = NULL,
  reopenedby = NULL, tOpen = NULL, request = NULL, tRequest = NULL,
  closedby = NULL, tClose = NULL, count = NULL)

CleanAll()

Clean(db)

CheckDB(db = NULL)

ConnString(db = NULL)

Databases()
}
\arguments{
\item{db}{A character value naming the database to use}
}
\description{
Functions to get/set connection objects and attributes.
}
\section{Functions}{
\itemize{
\item \code{Databases}: An internal function to used by connection functions that provide
all the information used to build the connection string. This is where new connections will be
configured
}}
\keyword{datasets}

