% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal_rodbcwrappers.R
\name{.odbcOpen}
\alias{.odbcClose}
\alias{.odbcOpen}
\alias{.odbcReOpen}
\alias{CheckDB}
\alias{Clean}
\alias{CleanAll}
\alias{ConnString}
\alias{Databases}
\alias{GetConn}
\alias{SetConn}
\title{Internal Connection Functions}
\usage{
.odbcOpen(db)

.odbcClose(db)

.odbcReOpen(db)

GetConn(db)

SetConn(object, name, status = NULL, openedby = NULL, reopenedby = NULL,
  tOpen = NULL, request = NULL, tRequest = NULL, closedby = NULL,
  tClose = NULL, count = NULL)

CleanAll()

Clean(db)

CheckDB(db = NULL)

ConnString(db = NULL)

Databases()
}
\arguments{
\item{db}{A character value naming the database to use}
}
\value{
Object of class RODBC or NULL in the case of .odbcClose()
}
\description{
These are functions that are not exposed to users. They are custom wrappers around connection functions from
the package RODBC. Additionally, they provide internal functionality to build connection strings, set and get
connection attributes, add connections to and clean up the connection pool, etc. These are abstracted from users/developers
in order to enable developers to efficiently build data products.
}
\section{Functions}{
\itemize{
\item \code{Databases}: An internal function to used by connection functions that provide
all the information used to build the connection string. This is where new connections will be
configured
}}
\section{Internal Use Only}{

 Internal functions not intended to be used by developers. To open and close connections to a preconfigured database,
 use: \code{\link{OpenDB}} or \code{\link{CloseDB}}.
}
\seealso{
RODBC Vignette: \url{https://cran.r-project.org/web/packages/RODBC/RODBC.pdf}
}

