% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-help_dbexplore.R
\name{TableInfo}
\alias{ColumnInfo}
\alias{Columns}
\alias{PrimaryKey}
\alias{TableInfo}
\alias{getColumns}
\alias{getPrimaryKey}
\alias{getTables}
\title{Tools to explore the tables or columns in a database}
\usage{
TableInfo(db = NULL, print = FALSE)

PrimaryKey(db = NULL, tables = NULL)

ColumnInfo(db = NULL, tables = NULL, closeConn = FALSE, print = FALSE)

Columns(db = NULL, tables = NULL, return = FALSE, closeConn = TRUE)

getTables(db = NULL)

getColumns(db = NULL, tables)

getPrimaryKey(db = NULL, tables)
}
\arguments{
\item{db}{A character value naming the database (e.g. "Morpheus")}

\item{print}{A boolean indicating whether to print information on the R console}

\item{tables}{A character vector of table names belonging to the database specified by "db"}

\item{closeConn}{A boolean indicating whether this function to close the connection to "db". Default is TRUE.}

\item{return}{A boolean indicating whether to return objects in memory. Defaults to FALSE which means
information is simply printed on the R console and not captured. Ideal default for quick lookups.}
}
\description{
Functions assist with exploring db objects at the abstraction layer, and general
exploration of the Morpheus data
}
\section{Functions}{
\itemize{
\item \code{TableInfo}: Given a database, get all databases on same server

\item \code{PrimaryKey}: A function to retrieve information about one or more table's primary key column

\item \code{ColumnInfo}: A function to detailed information about a given table's columns.
If argument "tables" not provided, retrive information about all tables in the specified database

\item \code{Columns}: A function that serves as a quick lookup of columns in a table. Defaults to table
"MainLog" in database "Morpheus".

\item \code{getTables}: An internal wrapper around sqlTables

\item \code{getColumns}: An internal wrapper around sqlColumns

\item \code{getPrimaryKey}: An internal wrapper around sqlPrimaryKeys
}}
\examples{
\dontrun{
# Let's explore what data is available in Morpheus

# Available tables?
TableInfo(db = "Morpheus")
TableInfo(db = "RemoteFE", print = T)

# What are the keys for each table?
PrimaryKey("Morpheus")

# Get key for a specific table
PrimaryKey("Morpheus", "MainLog")

# Let's explore MainLog columns
ColumnInfo(tables = "MainLog")

# Shortcuts/quick view
Columns()
Columns(table = "ProcedureSummaries")
}
}

